feat(button): render a react-router link
A button can have a `href` attribute but it will always do a full page reload when clicked. This
makes it possible to pass in a `to` attribute which will make it a react-router link.

BREAKING CHANGE: react-router-dom is a peerDepencency now. All packages using om will have to
install react-router-dom too.